+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 2.1.1 under development
 
-- no changes in this release.
+- Enh #121: Add support of default value when getting non-exist object property value (@vjik) 
 
 ## 2.1.0 August 20, 2022